+++ src/node/mod.rs
@@ -27,8 +27,8 @@ mod panner;
 pub use panner::*;
 mod analyzer;
 pub use analyzer::*;
-mod audio_buffer_src;
-pub use audio_buffer_src::*;
+mod audio_buffer;
+pub use audio_buffer::*;
 mod media_element;
 pub use media_element::*;
 mod media_stream;
